requirements.txt
fastapi==0.111.0
httpx==0.27.0
sqlmodel==0.0.19
psycopg2-binary==2.9.9
python-dotenv==1.0.1